fix: ensure consistent theme detection between popup and dashboard
This commit addresses the inconsistency between system theme detection in the
popup and dashboard contexts. The popup was detecting a dark theme preference
while the dashboard correctly detected a light theme preference.

Changes:
- Create shared theme detection helper in src/lib/themeHelper.js
- Add robust system preference detection that works across contexts
- Add theme-init.css to popup for consistent theme flash prevention
- Update both popup and dashboard main.js to use the shared helper
- Add proper transition handling to prevent UI flashing in popup

This ensures a consistent theming experience across all parts of the extension,
fixing the inconsistency between window.matchMedia results in different contexts.
