+++ src/popup/popup.html
@@ -4,6 +4,8 @@
   <meta charset="UTF-8">
   <meta name="viewport" content="width=device-width, initial-scale=1.0">
   <title>Gemini History Manager</title>
+  <!-- Critical CSS for theme initialization to prevent flash -->
+  <link rel="stylesheet" href="theme-init.css">
   <link rel="stylesheet" href="popup.css">
 </head>
 <body>
