+++ src/popup/theme-init.css
@@ -1,0 +1,31 @@
+/* 
+ * Theme initialization CSS for popup
+ * Critical CSS loaded before the main stylesheet to prevent theme flash
+ */
+
+/* Default light theme background */
+html {
+  background-color: #f9f9fb; 
+}
+
+/* Explicit dark theme */
+html[data-theme='dark'] {
+  background-color: #121212; 
+}
+
+/* Explicit light theme - must override system preference */
+html[data-theme='light'] {
+  background-color: #f9f9fb !important;
+}
+
+/* System preference dark mode - for when the theme attribute isn't set yet */
+@media (prefers-color-scheme: dark) {
+  html:not([data-theme]) {
+    background-color: #121212;
+  }
+}
+
+/* Disable transitions during initial load to prevent flickering */
+.initial-load * {
+  transition: none !important;
+}
